+++ taxonkit/cmd/reformat.go
@@ -216,12 +216,8 @@ Output format can contains some escape charactors like "\t".
 
 		blankS := format
 		iblankS := format
-		for _, re := range reRankPlaceHolders {
-			blankS = re.ReplaceAllString(blankS, blank)
-		}
-		for _, re := range reRankPlaceHolders {
-			iblankS = re.ReplaceAllString(iblankS, iblank)
-		}
+		blankS = reRankPlaceHolder.ReplaceAllString(blankS, blank)
+		iblankS = reRankPlaceHolder.ReplaceAllString(iblankS, iblank)
 
 		fn := func(line string) (interface{}, bool, error) {
 			if len(line) == 0 || line[0] == '#' {