+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+- [TaxonKit v0.19.1](https://github.com/shenwei356/taxonkit/releases/tag/v0.19.1)
+[![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/taxonkit/v0.19.1/total.svg)](https://github.com/shenwei356/taxonkit/releases/tag/v0.19.1)
+    - `taxonkit reformat2`:
+        - Fix handling deleted TaxIds. [#117](https://github.com/shenwei356/taxonkit/issues/117)
 - [TaxonKit v0.19.0](https://github.com/shenwei356/taxonkit/releases/tag/v0.19.0)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/taxonkit/v0.19.0/total.svg)](https://github.com/shenwei356/taxonkit/releases/tag/v0.19.0)
     - new command `taxonkit reformat2`: Reformat lineage in chosen ranks, allowing more ranks than 'reformat'